### 2019-08-01 #1615 @labarba
@liashchynskyi — I am the Associate-Editor-in-Chief on rotation, and will be looking for a handling editor for your submission.

After reading your paper, though, it seems like a bit of a slog to have to read a whole page of background before getting to a statement about what the software does. Could you look into placing a few sentences about the software right at the top of your summary? The reader wants to know at the start: _Who is this software for? What does it do?_

You will also have to work on the figure placements. Once you make changes, you can run the command `@whedon generate pdf` right here in a comment to get a new proof.
