### 2019-08-01 #1615 @labarba
Our [Documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) requests:

> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
