### 2019-08-01 5226c16
Update paper.md
