### 2019-08-01 2b89147
Update paper.md
