+++ paper/paper.md
@@ -21,7 +21,11 @@ bibliography: refs.bib
 
 # Purpose of the software
 
-I present Rudi, a Python tool that simplifies the dataset preparation for training a convolutional neural network or other things that need images as input data. The tool contains a set of practical generalized tools, which can wrap common operations like image resizing, flipping, zooming, distortion and others. Presented tool can be helpful with extending the dateset when the user does not have enough of data.
+I present Rudi, a Python tool that simplifies the dataset preparation for training a convolutional neural network or other things that need images as input data. 
+The tool contains a set of practical generalized tools, which can wrap common operations like image resizing, flipping, zooming, distortion and others. 
+Presented tool can be helpful with extending the dateset when the user does not have enough of data. After installation `Rudi`, the tool can be called by typing `rudi` in the terminal window (see Figure 1).
+
+![Output information after calling `Rudi` script in terminal. ](figure1.png)
 
 # Background
 
@@ -68,10 +72,6 @@ and data augmentation. [^1] [^2]
 
 Rudi is implemented in Python. It takes as input a root directory of dataset images and a few additional 
 options and outputs converted images to *output* directory in a root. 
-
-After installation `Rudi`, the tool can be called by typing `rudi` in the terminal window (see Figure 1).
-
-![Output information after calling `Rudi` script in terminal. ](figure1.png)
 
 Rudi tool provides two main commands:
 
