+++ paper/paper.md
@@ -23,9 +23,7 @@ bibliography: refs.bib
 
 I present Rudi, a Python tool that simplifies the dataset preparation for training a convolutional neural network or other things that need images as input data. 
 The tool contains a set of practical generalized tools, which can wrap common operations like image resizing, flipping, zooming, distortion and others. 
-Presented tool can be helpful with extending the dateset when the user does not have enough of data. After installation `Rudi`, the tool can be called by typing `rudi` in the terminal window (see Figure 1).
-
-![Output information after calling `Rudi` script in terminal. ](figure1.png)
+Presented tool can be helpful with extending the dateset when the user does not have enough of data. After installation `Rudi`, the tool can be called by typing `rudi` in the terminal window.
 
 # Background
 
@@ -89,11 +87,11 @@ rotating, zooming, random distortion and skewing.
 
 # Examples
 
-This tool was used for expanding the dataset of biomedical images [@tneufcit2019] from under 100 to more than 500 images (see Figure 2).
+This tool was used for expanding the dataset of biomedical images [@tneufcit2019] from under 100 to more than 500 images (see Figure 1).
 
 ![Example images of the original and augmented dataset. ](figure2.png)
 
-As described by @tneufcit2019 this approach allowed author to achieve *95%* accuracy on *histological images dataset* (5 classes) (see Figure 3).
+As described by @tneufcit2019 this approach allowed author to achieve *95%* accuracy on *histological images dataset* (5 classes) (see Figure 2).
 
 ![ROC crurve for CNN classifier for histological images. ](figure3.png)
 
