### 2019-08-03 #1615 @labarba
👋  @mbobra — Could you handle this new JOSS submission as editor?

**Rudi: Lightweight image converter and dataset augmentor**
https://github.com/liashchynskyi/rudi
[ Article proof :page_facing_up: ](https://github.com/openjournals/joss-papers/blob/joss.01615/joss.01615/10.21105.joss.01615.pdf)
