### 2019-08-07 #1615 @arfon
@liashchynskyi - thanks for your submission. Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
