### 2019-08-07 #1615 @arfon
@shivam11 @soroushj @casperdcl would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
