### 2019-08-19 #1615 @arfon
@clemaitre58 @nnadeau - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
