### 2019-08-20 #1664 @whedon
[REVIEW]: Rudi: Lightweight image converter and dataset augmentor
**Submitting author:** @liashchynskyi (<a href="http://orcid.org/0000-0002-3920-6239">Petro Liashchynskyi</a>)
**Repository:** <a href="https://github.com/liashchynskyi/rudi" target ="_blank">https://github.com/liashchynskyi/rudi</a>
**Version:** 1.0
