### 2019-08-20 #1664 @casperdcl
# Review

Thought I'd be über quick about this, considering how small the work is it was very easy to review.

## A statement of need

- https://github.com/liashchynskyi/rudi/blame/88b3ed52ef557cba9f2f4db9280c861caa0b86d1/paper/paper.md#L39-L566: Excessive repetition defining the concept of "data augmentation" and justifying its existence, which doesn't seem to a point which needs debate.
- The package (*Rudi*) lists *Augmentor* (https://github.com/mdbloice/Augmentor, http://augmentor.readthedocs.io/) as a requirement. It would appear that *Augmentor*'s statement of need is essentially identical to Rudi's. This, combined with the fact that *Rudi* is an extremely tiny wrapper package (115 lines of python code excluding `setup.py` and blanks, and only 76 if counting `augmentor.py` and `converter.py`) with a very short development history (~1 month) makes it seem far better placed to be a pull request on *Augmentor* rather than a stand-alone package.

## Installation instructions

- Dependencies are only listed in `setup.py`. Ideally they should also appear in `README.md` and/or `requirements.txt`.

## Functionality documentation

- Currently only provided by installing and running help on the executable.

## Automated tests

- There are no tests.

## Community guidelines

- There are no guidelines.

## Overall recommendation

- Convert into a small sub-package of https://github.com/mdbloice/Augmentor and open a PR there
