### 2019-08-20 #1664 @arfon
Thanks for your review @casperdcl. It sounds like this package might fall foul of our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) that explicitly call out 'minor utilities' as not allowable in JOSS:

> - The software should be feature-complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications). Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable.

In your opinion does this package fall into that category?
