### 2019-08-21 #1664 @casperdcl
@arfon yes, while it may be a useful package it really does seem to be a thin API wrapper that could be present in *Augmentor*.
