### 2019-08-21 #1664 @arfon
OK thanks @casperdcl.

@nnadeau - could you give your opinion on [this question](https://github.com/openjournals/joss-reviews/issues/1664#issuecomment-523236126) too?
