### 2019-08-21 #1664 @casperdcl
@liashchynskyi I don't have a problem with using dependencies (things like [nteract/papermill](https://github.com/nteract/papermill) even recommend you [extend them by only writing a minimal `setup.py` file and nothing else!](https://papermill.readthedocs.io/en/latest/extending-entry-points.html)), it's just as @arfon says not sure if it's appropriate for a JOSS publication.
