### 2019-08-21 #1664 @liashchynskyi
ok, close the issue 😁 good luck, guys

Ср, 21 серп. 2019 16:06 користувач Casper da Costa-Luis <
notifications@github.com> пише:

> @liashchynskyi <https://github.com/liashchynskyi> I don't have a problem
> with using dependencies (things like nteract/papermill
> <https://github.com/nteract/papermill> even recommend you extend them by
> only writing a minimal setup.py file and nothing else!
> <https://papermill.readthedocs.io/en/latest/extending-entry-points.html>),
> it's just as @arfon <https://github.com/arfon> says not sure if it's
> appropriate for a JOSS publication.
>
