### 2019-08-22 #1664 @arfon
OK thanks for your reviews here @casperdcl & @nnadeau

@liashchynskyi - we'd welcome a resubmission in the future if this package were extended/made more comprehensive such that it met our [submission criteria](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements).
